src/main/java/org/intro/dynamicConnectivityProblem/LazyQuickUnion.java
package org.intro.dynamicConnectivityProblem;

public class LazyQuickUnion {

    private final int[] id;

    public LazyQuickUnion(int n) {
        id = new int[n];
        for (int i = 0; i < n; i++) {
            id[i] = i;
        }
    }

    public int root(int i) {
        while (i != id[i]) {
            i = id[i];
        }
        return i;
    }

    public boolean connected(int p, int q) {
        return root(p) == root(q);
    }

    public void union(int p, int q) {
        int pid = root(p);
        int qid = root(q);
        id[pid] = qid;
    }
}
